Filter out landing pages from the S3 data
